+++ codespan-reporting/src/files.rs
@@ -25,6 +25,58 @@
 
 use std::ops::Range;
 
+/// An enum representing an error that happened while looking up a file or a piece of content in that file.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum Error {
+    /// A required file is not in the file database.
+    FileMissing,
+    /// The file is present, but does not contain the specified byte index.
+    IndexTooLarge { given: usize, max: usize },
+    /// The file is present, but does not contain the specified line index.
+    LineTooLarge { given: usize, max: usize },
+    /// The file is present and contains the specified line index, but the line does not contain the specified column index.
+    ColumnTooLarge { given: usize, max: usize },
+    /// The given index is contained in the file, but is not a boundary of a UTF-8 code point.
+    InvalidCharBoundary { given: usize },
+    /// There was a error while doing IO.
+    Io(std::io::Error),
+}
+
+impl From<std::io::Error> for Error {
+    fn from(err: std::io::Error) -> Error {
+        Error::Io(err)
+    }
+}
+
+impl std::fmt::Display for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Error::FileMissing => write!(f, "file missing"),
+            Error::IndexTooLarge { given, max } => {
+                write!(f, "invalid index {}, maximum index is {}", given, max)
+            }
+            Error::LineTooLarge { given, max } => {
+                write!(f, "invalid line {}, maximum line is {}", given, max)
+            }
+            Error::ColumnTooLarge { given, max } => {
+                write!(f, "invalid column {}, maximum column {}", given, max)
+            }
+            Error::InvalidCharBoundary { .. } => write!(f, "index is not a code point boundary"),
+            Error::Io(err) => write!(f, "{}", err),
+        }
+    }
+}
+
+impl std::error::Error for Error {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match &self {
+            Error::Io(err) => Some(err),
+            _ => None,
+        }
+    }
+}
+
 /// A minimal interface for accessing source files when rendering diagnostics.
 ///
 /// A lifetime parameter `'a` is provided to allow any of the returned values to returned by reference.
@@ -40,10 +92,10 @@ pub trait Files<'a> {
     type Source: 'a + AsRef<str>;
 
     /// The user-facing name of a file.
-    fn name(&'a self, id: Self::FileId) -> Option<Self::Name>;
+    fn name(&'a self, id: Self::FileId) -> Result<Self::Name, Error>;
 
     /// The source code of a file.
-    fn source(&'a self, id: Self::FileId) -> Option<Self::Source>;
+    fn source(&'a self, id: Self::FileId) -> Result<Self::Source, Error>;
 
     /// The index of the line at the given byte index.
     /// If the byte index is at or past the end of the file, returns the maximum line index in the file.
@@ -58,7 +110,7 @@ pub trait Files<'a> {
     ///
     /// [`line_starts`]: crate::files::line_starts
     /// [`files`]: crate::files
-    fn line_index(&'a self, id: Self::FileId, byte_index: usize) -> Option<usize>;
+    fn line_index(&'a self, id: Self::FileId, byte_index: usize) -> Result<usize, Error>;
 
     /// The user-facing line number at the given line index.
     /// It is not necessarily checked that the specified line index
@@ -72,8 +124,8 @@ pub trait Files<'a> {
     ///
     /// [line-macro]: https://en.cppreference.com/w/c/preprocessor/line
     #[allow(unused_variables)]
-    fn line_number(&'a self, id: Self::FileId, line_index: usize) -> Option<usize> {
-        Some(line_index + 1)
+    fn line_number(&'a self, id: Self::FileId, line_index: usize) -> Result<usize, Error> {
+        Ok(line_index + 1)
     }
 
     /// The user-facing column number at the given line index and byte index.
@@ -91,27 +143,27 @@ pub trait Files<'a> {
         id: Self::FileId,
         line_index: usize,
         byte_index: usize,
-    ) -> Option<usize> {
+    ) -> Result<usize, Error> {
         let source = self.source(id)?;
         let line_range = self.line_range(id, line_index)?;
         let column_index = column_index(source.as_ref(), line_range, byte_index);
 
-        Some(column_index + 1)
+        Ok(column_index + 1)
     }
 
     /// Convenience method for returning line and column number at the given
     /// byte index in the file.
-    fn location(&'a self, id: Self::FileId, byte_index: usize) -> Option<Location> {
+    fn location(&'a self, id: Self::FileId, byte_index: usize) -> Result<Location, Error> {
         let line_index = self.line_index(id, byte_index)?;
 
-        Some(Location {
+        Ok(Location {
             line_number: self.line_number(id, line_index)?,
             column_number: self.column_number(id, line_index, byte_index)?,
         })
     }
 
     /// The byte range of line in the source of the file.
-    fn line_range(&'a self, id: Self::FileId, line_index: usize) -> Option<Range<usize>>;
+    fn line_range(&'a self, id: Self::FileId, line_index: usize) -> Result<Range<usize>, Error>;
 }
 
 /// A user-facing location in a source file.
@@ -259,27 +311,41 @@ where
     type Name = Name;
     type Source = &'a str;
 
-    fn name(&self, (): ()) -> Option<Name> {
-        Some(self.name.clone())
+    fn name(&self, (): ()) -> Result<Name, Error> {
+        Ok(self.name.clone())
     }
 
-    fn source(&self, (): ()) -> Option<&str> {
-        Some(self.source.as_ref())
+    fn source(&self, (): ()) -> Result<&str, Error> {
+        Ok(self.source.as_ref())
     }
 
-    fn line_index(&self, (): (), byte_index: usize) -> Option<usize> {
-        match self.line_starts.binary_search(&byte_index) {
-            // never return the last line index, because that is not a "real" line
-            Ok(line) => Some(std::cmp::min(line, self.line_starts.len() - 2)),
-            Err(next_line) => Some(std::cmp::min(next_line - 1, self.line_starts.len() - 2)),
-        }
+    fn line_index(&self, (): (), byte_index: usize) -> Result<usize, Error> {
+        Ok(self
+            .line_starts
+            .binary_search(&byte_index)
+            .map(|line| std::cmp::min(line, self.line_starts.len() - 2))
+            .unwrap_or_else(|next_line| std::cmp::min(next_line - 1, self.line_starts.len() - 2)))
     }
 
-    fn line_range(&self, (): (), line_index: usize) -> Option<Range<usize>> {
-        let line_start = self.line_starts.get(line_index).copied()?;
-        let next_line_start = self.line_starts.get(line_index + 1).copied()?;
+    fn line_range(&self, (): (), line_index: usize) -> Result<Range<usize>, Error> {
+        let line_start = self
+            .line_starts
+            .get(line_index)
+            .copied()
+            .ok_or(Error::LineTooLarge {
+                given: line_index,
+                max: self.line_starts.len() - 1,
+            })?;
+        let next_line_start =
+            self.line_starts
+                .get(line_index + 1)
+                .copied()
+                .ok_or(Error::LineTooLarge {
+                    given: line_index,
+                    max: self.line_starts.len() - 1,
+                })?;
 
-        Some(line_start..next_line_start)
+        Ok(line_start..next_line_start)
     }
 }
 
@@ -312,8 +378,8 @@ where
     }
 
     /// Get the file corresponding to the given id.
-    pub fn get(&self, file_id: usize) -> Option<&SimpleFile<Name, Source>> {
-        self.files.get(file_id)
+    pub fn get(&self, file_id: usize) -> Result<&SimpleFile<Name, Source>, Error> {
+        self.files.get(file_id).ok_or(Error::FileMissing)
     }
 }
 
@@ -326,19 +392,19 @@ where
     type Name = Name;
     type Source = &'a str;
 
-    fn name(&self, file_id: usize) -> Option<Name> {
-        Some(self.get(file_id)?.name().clone())
+    fn name(&self, file_id: usize) -> Result<Name, Error> {
+        Ok(self.get(file_id)?.name().clone())
     }
 
-    fn source(&self, file_id: usize) -> Option<&str> {
-        Some(self.get(file_id)?.source().as_ref())
+    fn source(&self, file_id: usize) -> Result<&str, Error> {
+        Ok(self.get(file_id)?.source().as_ref())
     }
 
-    fn line_index(&self, file_id: usize, byte_index: usize) -> Option<usize> {
+    fn line_index(&self, file_id: usize, byte_index: usize) -> Result<usize, Error> {
         self.get(file_id)?.line_index((), byte_index)
     }
 
-    fn line_range(&self, file_id: usize, line_index: usize) -> Option<Range<usize>> {
+    fn line_range(&self, file_id: usize, line_index: usize) -> Result<Range<usize>, Error> {
         self.get(file_id)?.line_range((), line_index)
     }
 }